+++ src/components/layer-controller/LayerController.js
@@ -1,41 +1,65 @@
 import React, { useState } from 'react';
-import { getChannelStats, MAX_SLIDERS_AND_CHANNELS } from '@hms-dbmi/viv';
+import { MAX_SLIDERS_AND_CHANNELS, getChannelStats } from '@hms-dbmi/viv';
 
 import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import AddIcon from '@material-ui/icons/Add';
 import Slider from '@material-ui/core/Slider';
-import InputLabel from '@material-ui/core/InputLabel';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
 
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+import VisibilityIcon from '@material-ui/icons/Visibility';
+import VisibilityOffIcon from '@material-ui/icons/VisibilityOff';
 
 import LayerOptions from './LayerOptions';
-
+import VolumeOptions from './VolumeOptions';
 import {
-  useExpansionPanelStyles,
-  useExpansionPanelSummaryStyles,
-  useSmallInputLabelStyles,
+  useControllerSectionStyles,
+  StyledExpansionPanelDetails,
+  StyledExpansionPanelSummary,
+  StyledInputLabel,
+  OverflowEllipsisGrid,
 } from './styles';
+import { getMultiSelectionStats } from './utils';
+
 import { GLOBAL_LABELS } from '../spatial/constants';
 import { getSourceFromLoader, isRgb } from '../../utils';
 import { DOMAINS } from './constants';
 
 
-// Set the domain of the sliders based on either a full range or min/max.
-async function getDomainsAndSliders(loader, loaderSelection, domainType) {
-  let domains;
-  const source = getSourceFromLoader(loader);
-  const raster = await Promise.all(
-    loaderSelection.map(selection => source.getRaster({ selection })),
+function TabPanel(props) {
+  const {
+    children, value, index, ...other
+  } = props;
+
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && children}
+    </div>
   );
-  const stats = raster.map(({ data: d }) => getChannelStats(d));
-  const sliders = stats.map(stat => stat.autoSliders);
+}
+
+
+// Set the domain of the sliders based on either a full range or min/max.
+async function getDomainsAndSliders(loader, loaderSelection, domainType, use3d) {
+  let domains;
+  const stats = await getMultiSelectionStats({
+    loader: loader.data, selections: loaderSelection, use3d,
+  });
+  const { sliders } = stats;
   if (domainType === 'Min/Max') {
-    domains = stats.map(stat => stat.domain);
+    // eslint-disable-next-line prefer-destructuring
+    domains = stats.domains;
   } if (domainType === 'Full') {
+    const source = getSourceFromLoader(loader);
     domains = loaderSelection.map(() => DOMAINS[source.dtype]);
   }
   return { domains, sliders };
@@ -57,6 +81,9 @@ export default function LayerController(props) {
     handleLayerRemove, handleLayerChange,
     shouldShowTransparentColor,
     shouldShowDomain, shouldShowColormap, ChannelController,
+    setViewState, disable3d, setRasterLayerCallback,
+    setAreLayerChannelsLoading, areLayerChannelsLoading, disabled,
+    spatialHeight, spatialWidth,
   } = props;
 
   const {
@@ -64,10 +91,22 @@ export default function LayerController(props) {
     opacity,
     channels,
     transparentColor,
+    renderingMode,
+    xSlice,
+    ySlice,
+    zSlice,
+    resolution,
+    use3d,
+    modelMatrix,
   } = layer;
   const firstSelection = channels[0]?.selection || {};
 
   const { data, channels: channelOptions } = loader;
+  const [tab, setTab] = useState(0);
+
+  const handleTabChange = (event, newTab) => {
+    setTab(newTab);
+  };
   const { labels, shape } = Array.isArray(data) ? data[data.length - 1] : data;
   const [domainType, setDomainType] = useState(layer.domainType);
   const [isExpanded, setIsExpanded] = useState(true);
@@ -76,6 +115,10 @@ export default function LayerController(props) {
       .filter(field => typeof firstSelection[field] === 'number')
       .reduce((o, key) => ({ ...o, [key]: firstSelection[key] }), {}),
   );
+
+  function setVisible(v) {
+    handleLayerChange({ ...layer, visible: v });
+  }
 
   function setColormap(v) {
     handleLayerChange({ ...layer, colormap: v });
@@ -90,6 +133,17 @@ export default function LayerController(props) {
   }
   function setTransparentColor(v) {
     handleLayerChange({ ...layer, transparentColor: v });
+  }
+  function setRenderingMode(v) {
+    handleLayerChange({ ...layer, renderingMode: v });
+  }
+
+  function handleMultiPropertyChange(obj) {
+    handleLayerChange({ ...layer, ...obj });
+  }
+
+  function handleSlicerSetting(slice, val) {
+    handleLayerChange({ ...layer, [`${slice}Slice`]: val });
   }
 
   function setChannelsAndDomainType(newChannels, newDomainType) {
@@ -117,6 +171,11 @@ export default function LayerController(props) {
     handleLayerChange({ ...layer, channels: newChannels });
   }
 
+  const setAreAllChannelsLoading = (val) => {
+    const newAreLayerChannelsLoading = channels.map(() => val);
+    setAreLayerChannelsLoading(newAreLayerChannelsLoading);
+  };
+
   // Handles adding a channel, creating a default selection
   // for the current global settings and domain type.
   const handleChannelAdd = async () => {
@@ -128,14 +187,26 @@ export default function LayerController(props) {
         ? (globalLabelValues[label] || 0)
         : 0;
     });
-    const { domains, sliders } = await getDomainsAndSliders(loader, [selection], domainType);
+    const { domains, sliders } = await getDomainsAndSliders(loader, [selection], domainType, use3d);
     const domain = domains[0];
-    const slider = sliders[0] || domain;
+    const slider = domain;
     const color = [255, 255, 255];
     const visible = true;
-    addChannel({
+    const newChannelId = channels.length;
+    const newAreLayerChannelsLoading = [...areLayerChannelsLoading];
+    newAreLayerChannelsLoading[newChannelId] = true;
+    setAreLayerChannelsLoading(newAreLayerChannelsLoading);
+    const channel = {
       selection, slider, visible, color,
+    };
+    setRasterLayerCallback(() => {
+      setChannel({ ...channel, slider: sliders[0] }, newChannelId);
+      const areLayerChannelsLoadingCallback = [...newAreLayerChannelsLoading];
+      areLayerChannelsLoadingCallback[newChannelId] = false;
+      setAreLayerChannelsLoading(areLayerChannelsLoadingCallback);
+      setRasterLayerCallback(null);
     });
+    addChannel(channel);
   };
 
   const handleDomainChange = async (value) => {
@@ -150,6 +221,7 @@ export default function LayerController(props) {
       loader,
       loaderSelection,
       value,
+      use3d,
     );
 
     // If it's the right-most slider, we take the minimum of that and the new value.
@@ -174,18 +246,30 @@ export default function LayerController(props) {
       ...channel.selection,
       ...selection,
     }));
-    const mouseUp = event.type === 'mouseup';
+    const canUpdateChannels = event.type === 'mouseup' || event.type === 'keydown';
     // Only update domains on a mouseup event for the same reason as above.
-    const { sliders } = mouseUp
-      ? await getDomainsAndSliders(loader, loaderSelection, domainType)
-      : { domains: [], sliders: [] };
-    if (mouseUp) {
-      const newChannels = channels.map((c, i) => ({
-        ...c,
-        slider: sliders[i],
-        selection: { ...c.selection, ...selection },
-      }));
-      setChannels(newChannels);
+    if (canUpdateChannels) {
+      setAreAllChannelsLoading(true);
+      getDomainsAndSliders(loader, loaderSelection, domainType, use3d).then(({ sliders }) => {
+        const newChannelsWithSelection = channels.map(c => ({
+          ...c,
+          selection: { ...c.selection, ...selection },
+        }));
+        // Set the callback before changing the selection
+        // so the callback is used when the layer (re)loads its data.
+        setRasterLayerCallback(() => {
+          setRasterLayerCallback(null);
+          setAreAllChannelsLoading(false);
+          const newChannelsWithSliders = [...newChannelsWithSelection].map(
+            (c, i) => ({
+              ...c,
+              slider: sliders[i],
+            }),
+          );
+          setChannels(newChannelsWithSliders);
+        });
+        setChannels(newChannelsWithSelection);
+      });
     }
     setGlobalLabelValues(prev => ({ ...prev, ...selection }));
   };
@@ -197,23 +281,44 @@ export default function LayerController(props) {
     channelControllers = channels.map(
       // c is an object like { color, selection, slider, visible }.
       (c, channelId) => {
+        // Update the auxiliary store with the current loading state of a channel.
+        const setIsLoading = (val) => {
+          const newAreLayerChannelsLoading = [...areLayerChannelsLoading];
+          newAreLayerChannelsLoading[channelId] = val;
+          setAreLayerChannelsLoading(newAreLayerChannelsLoading);
+        };
         // Change one property of a channel (for now - soon
         // nested structures allowing for multiple z/t selecitons at once, for example).
-        const handleChannelPropertyChange = async (property, value) => {
+        const handleChannelPropertyChange = (property, value) => {
           // property is something like "selection" or "slider."
           // value is the actual change, like { channel: "DAPI" }.
           const update = { [property]: value };
           if (property === 'selection') {
-            update.selection = { ...globalLabelValues, ...update.selection };
-            const loaderSelection = [
-              { ...channels[channelId][property], ...value },
-            ];
-            const { sliders } = await getDomainsAndSliders(
-              loader, loaderSelection, domainType,
-            );
-            [update.slider] = sliders;
+            // Channel is loading until the layer callback is called
+            // by the layer, which fetches the raster data.
+            setIsLoading(true);
+            update.selection = {
+              ...globalLabelValues,
+              ...update.selection,
+            };
+            setChannel({ ...c, ...update }, channelId);
+            // Call back for raster layer handles update of UI
+            // like sliders and the loading state of the channel.
+            setRasterLayerCallback(async () => {
+              const loaderSelection = [
+                { ...channels[channelId][property], ...value },
+              ];
+              const { sliders } = await getDomainsAndSliders(
+                loader, loaderSelection, domainType, use3d,
+              );
+              [update.slider] = sliders;
+              setChannel({ ...c, ...update }, channelId);
+              setRasterLayerCallback(null);
+              setIsLoading(false);
+            });
+          } else {
+            setChannel({ ...c, ...update }, channelId);
           }
-          setChannel({ ...c, ...update }, channelId);
         };
         const handleChannelRemove = () => {
           removeChannel(channelId);
@@ -248,34 +353,103 @@ export default function LayerController(props) {
             handlePropertyChange={handleChannelPropertyChange}
             handleChannelRemove={handleChannelRemove}
             handleIQRUpdate={handleIQRUpdate}
+            setRasterLayerCallback={setRasterLayerCallback}
+            isLoading={areLayerChannelsLoading[channelId]}
+            use3d={use3d}
           />
         );
       },
     );
   }
 
-  const classes = useExpansionPanelStyles();
-  const summaryClasses = useExpansionPanelSummaryStyles();
-  const closedOpacityLabelClasses = useSmallInputLabelStyles();
+  const controllerSectionClasses = useControllerSectionStyles();
 
+  const { visible } = layer;
+  const Visibility = visible ? VisibilityIcon : VisibilityOffIcon;
+  // Only show Volume tabs if 3D is available.
+  const useVolumeTabs = !disable3d && shape[labels.indexOf('z')] > 1;
+  const FullController = (
+    <>
+      <LayerOptions
+        channels={channels}
+        opacity={opacity}
+        colormap={colormap}
+        transparentColor={transparentColor}
+        domainType={domainType}
+      // Only allow for global dimension controllers that
+      // exist in the `dimensions` part of the loader.
+        globalControlLabels={labels.filter(label => GLOBAL_LABELS.includes(label))}
+        globalLabelValues={globalLabelValues}
+        handleOpacityChange={setOpacity}
+        handleColormapChange={setColormap}
+        handleGlobalChannelsSelectionChange={handleGlobalChannelsSelectionChange}
+        handleTransparentColorChange={setTransparentColor}
+        isRgb={isRgb(loader)}
+        handleDomainChange={handleDomainChange}
+        shouldShowTransparentColor={shouldShowTransparentColor}
+        shouldShowDomain={shouldShowDomain}
+        shouldShowColormap={shouldShowColormap}
+        use3d={use3d}
+        loader={loader}
+        handleMultiPropertyChange={handleMultiPropertyChange}
+        resolution={resolution}
+        disable3d={disable3d}
+        setRasterLayerCallback={setRasterLayerCallback}
+        setAreAllChannelsLoading={setAreAllChannelsLoading}
+        setViewState={setViewState}
+        spatialHeight={spatialHeight}
+        spatialWidth={spatialWidth}
+        modelMatrix={modelMatrix}
+      />
+      {!isRgb(loader) ? channelControllers : null}
+      {!isRgb(loader) && (
+      <Button
+        disabled={channels.length === MAX_SLIDERS_AND_CHANNELS}
+        onClick={handleChannelAdd}
+        fullWidth
+        variant="outlined"
+        style={buttonStyles}
+        startIcon={<AddIcon />}
+        size="small"
+      >
+                Add Channel
+      </Button>
+      )}
+    </>
+  );
   return (
     <ExpansionPanel
-      className={classes.root}
-      onChange={(e, expanded) => setIsExpanded(expanded && e?.target?.attributes?.role?.value === 'presentation')}
+      className={controllerSectionClasses.root}
+      onChange={(e, expanded) => !disabled && setIsExpanded(expanded && e?.target?.attributes?.role?.value === 'presentation')}
       TransitionProps={{ enter: false }}
-      expanded={isExpanded}
+      expanded={!disabled && isExpanded}
     >
-      <ExpansionPanelSummary
+      <StyledExpansionPanelSummary
         expandIcon={<ExpandMoreIcon />}
         aria-controls={`layer-${name}-controls`}
-        classes={{ ...summaryClasses }}
       >
         <Grid container direction="column" m={1} justify="center">
-          <Grid item>{name}</Grid>
-          {!isExpanded && (
+          <OverflowEllipsisGrid item>
+            <Button
+              onClick={(e) => {
+                if (!disabled) {
+                  // Needed to prevent affecting the expansion panel from changing
+                  e.stopPropagation();
+                  setVisible(!visible);
+                }
+              }}
+              style={{
+                marginRight: 8, marginBottom: 2, padding: 0, minWidth: 0,
+              }}
+            >
+              <Visibility />
+            </Button>
+            {name}
+          </OverflowEllipsisGrid>
+          {!disabled && !isExpanded && !use3d && (
             <Grid container direction="row" alignItems="center" justify="center">
               <Grid item xs={6}>
-                <InputLabel htmlFor={`layer-${name}-opacity-closed`} classes={closedOpacityLabelClasses}>Opacity:</InputLabel>
+                <StyledInputLabel htmlFor={`layer-${name}-opacity-closed`}>Opacity:</StyledInputLabel>
               </Grid>
               <Grid item xs={6}>
                 <Slider
@@ -293,46 +467,57 @@ export default function LayerController(props) {
             </Grid>
           )}
         </Grid>
-      </ExpansionPanelSummary>
-      <ExpansionPanelDetails className={classes.root}>
-        <LayerOptions
-          channels={channels}
-          labels={labels}
-          shape={shape}
-          opacity={opacity}
-          colormap={colormap}
-          transparentColor={transparentColor}
-          domainType={domainType}
-            // Only allow for global dimension controllers that
-            // exist in the `dimensions` part of the loader.
-          globalControlLabels={labels.filter(label => GLOBAL_LABELS.includes(label))}
-          globalLabelValues={globalLabelValues}
-          handleOpacityChange={setOpacity}
-          handleColormapChange={setColormap}
-          handleGlobalChannelsSelectionChange={
-              handleGlobalChannelsSelectionChange
-            }
-          handleTransparentColorChange={setTransparentColor}
-          isRgb={isRgb(loader)}
-          handleDomainChange={handleDomainChange}
-          shouldShowTransparentColor={shouldShowTransparentColor}
-          shouldShowDomain={shouldShowDomain}
-          shouldShowColormap={shouldShowColormap}
-        />
-        {!isRgb(loader) ? channelControllers : null}
-        {!isRgb(loader) && (
-        <Button
-          disabled={channels.length === MAX_SLIDERS_AND_CHANNELS}
-          onClick={handleChannelAdd}
-          fullWidth
-          variant="outlined"
-          style={buttonStyles}
-          startIcon={<AddIcon />}
-          size="small"
-        >
-              Add Channel
-        </Button>
-        )}
+      </StyledExpansionPanelSummary>
+      <StyledExpansionPanelDetails>
+        {useVolumeTabs ? (
+          <>
+            <Tabs
+              value={tab}
+              onChange={handleTabChange}
+              aria-label="simple tabs example"
+              style={{ height: '24px', minHeight: '24px' }}
+            >
+              <Tab
+                label="Channels"
+                style={{
+                  fontSize: '.75rem',
+                  bottom: 12,
+                  width: '50%',
+                  minWidth: '50%',
+                }}
+                disableRipple
+              />
+              <Tab
+                label="Volume"
+                style={{
+                  fontSize: '.75rem',
+                  bottom: 12,
+                  width: '50%',
+                  minWidth: '50%',
+                }}
+              />
+            </Tabs>
+            <TabPanel value={tab} index={0}>
+              {FullController}
+            </TabPanel>
+            <TabPanel value={tab} index={1} style={{ marginTop: 4 }}>
+              <VolumeOptions
+                loader={loader}
+                handleSlicerSetting={handleSlicerSetting}
+                handleRenderingModeChange={setRenderingMode}
+                renderingMode={renderingMode}
+                xSlice={xSlice}
+                ySlice={ySlice}
+                zSlice={zSlice}
+                use3d={use3d}
+                setViewState={setViewState}
+                spatialHeight={spatialHeight}
+                spatialWidth={spatialWidth}
+                modelMatrix={modelMatrix}
+              />
+            </TabPanel>
+          </>
+        ) : FullController}
         <Button
           onClick={handleLayerRemove}
           fullWidth
@@ -340,9 +525,9 @@ export default function LayerController(props) {
           style={buttonStyles}
           size="small"
         >
-            Remove Image Layer
+          Remove Image Layer
         </Button>
-      </ExpansionPanelDetails>
+      </StyledExpansionPanelDetails>
     </ExpansionPanel>
   );
 }