+++ src/components/spatial/constants.js
@@ -1,12 +1,17 @@
+import { RENDERING_MODES } from '@hms-dbmi/viv';
+
 export const GLOBAL_LABELS = ['z', 't'];
 
 export const DEFAULT_RASTER_DOMAIN_TYPE = 'Min/Max';
 
 export const DEFAULT_RASTER_LAYER_PROPS = {
+  visible: true,
   colormap: null,
   opacity: 1,
   domainType: DEFAULT_RASTER_DOMAIN_TYPE,
   transparentColor: [0, 0, 0],
+  renderingMode: RENDERING_MODES.ADDITIVE,
+  use3d: false,
 };
 
 export const DEFAULT_MOLECULES_LAYER = {